+++ src/Graphics/AddScreen.java
@@ -1,3 +1,5 @@
+package Graphics;
+
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ActionEvent;
