+++ src/Graphics/RemoveScreen.java
@@ -1,15 +1,22 @@
-import api.DirectedWeightedGraphAlgorithms;
+package Graphics;
 
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
-public class ShortPathScreen {
+public class RemoveScreen {
 
+    enum Mode{
+        NODE,
+        EDGE
+    }
+
+    RemoveScreen.Mode screen;
     GUI gui;
 
-    public ShortPathScreen(GUI g) {
+    public RemoveScreen(GUI g, RemoveScreen.Mode mode) {
+        this.screen = mode;
         this.gui = g;
     }
 
@@ -29,6 +36,49 @@ public class ShortPathScreen {
         Header.setFont(new Font("Serif", Font.BOLD, 18));
         contentPane.add(Header);
 
+        if(screen == RemoveScreen.Mode.NODE){
+            RemoveNode(contentPane, frame);
+        }
+        else if(screen == RemoveScreen.Mode.EDGE){
+            RemoveEdge(contentPane, frame);
+        }
+
+    }
+
+    public void RemoveNode(Container contentPane, JFrame frame){
+        JTextArea key = new JTextArea("1");
+        key.setLocation(10, 100);
+        key.setSize(250, 50);
+        contentPane.add(key);
+
+        JButton button = new JButton("Remove Node");
+        button.setLocation(260, 100);
+        button.setSize(160, 50);
+        contentPane.add(button);
+
+        JLabel srcText = new JLabel("Node Key");
+        srcText.setBounds(10,75, 250,25);
+        srcText.setHorizontalAlignment(JTextField.CENTER);
+        srcText.setForeground(new Color(0, 0, 0));
+        srcText.setFont(new Font("Serif", Font.BOLD, 18));
+        contentPane.add(srcText);
+
+        button.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                String sNode = key.getText();
+                int keyVal = Integer.parseInt(sNode);
+
+                gui.removeNode(keyVal);
+
+                frame.dispose();
+
+            }
+        });
+
+        frame.setVisible(true);
+    }
+    public void RemoveEdge(Container contentPane, JFrame frame){
         JTextArea src = new JTextArea("1");
         src.setLocation(10, 100);
         src.setSize(125, 50);
@@ -39,19 +89,19 @@ public class ShortPathScreen {
         dst.setSize(120, 50);
         contentPane.add(dst);
 
-        JButton button = new JButton("Get Shortest Path");
+        JButton button = new JButton("Remove Node");
         button.setLocation(260, 100);
         button.setSize(160, 50);
         contentPane.add(button);
 
-        JLabel srcText = new JLabel("Source");
+        JLabel srcText = new JLabel("Source Key");
         srcText.setBounds(10,75, 125,25);
         srcText.setHorizontalAlignment(JTextField.CENTER);
         srcText.setForeground(new Color(0, 0, 0));
         srcText.setFont(new Font("Serif", Font.BOLD, 18));
         contentPane.add(srcText);
 
-        JLabel dstText = new JLabel("Destination");
+        JLabel dstText = new JLabel("Destination Key");
         dstText.setBounds(140,75, 120,25);
         dstText.setHorizontalAlignment(JTextField.CENTER);
         dstText.setForeground(new Color(0, 0, 0));
@@ -67,7 +117,7 @@ public class ShortPathScreen {
                 int srcKey = Integer.parseInt(srcNode);
                 int dstKey = Integer.parseInt(dstNode);
 
-                gui.shortestPath(srcKey, dstKey);
+                gui.removeEdge(srcKey, dstKey);
 
                 frame.dispose();
 
@@ -75,7 +125,5 @@ public class ShortPathScreen {
         });
 
         frame.setVisible(true);
-
     }
-
 }
